fix: use useAppStateListener hook and move it to src level
