+++ package/src/index.ts
@@ -1,4 +1,5 @@
 export * from './components';
+export * from './hooks';
 export { registerNativeHandlers, NetInfo } from './native';
 export * from './contexts';
 export * from './emoji-data/compiled';
